Various small fixes related to consequence calling
modules/Bio/EnsEMBL/Variation/BaseTranscriptVariation.pm
==
Fix to _overlapped_exons
The code for handling transcripts with frameshift introns was causing 'exon' pre to sometimes be
added erroneously, which caused misses of polypyrimidine consequences in an intron.
The 'exon' pre was set for variant which lay inside a real intron of a transcript which also
had frameshift introns because the 12 bp 'stretch' was applied to all vfs in the transcript.
Need to test if the exon is next to a frameshift intron which is part of the overlap.

Fix to _create_intron_trees
Changed bounds for some ranges added to intron boundary tree because for very short
(eg 1 bp frameshift intron) introns intron_end-8 will extend back BEFORE the intron and
even before it's splice region, leading to incorrect splice_region_variant calls

modules/Bio/EnsEMBL/Variation/BaseTranscriptVariationAllele.pm
==
Fixes to _intron_effects
Change to low - high rather than start - end for checking polypyrimidine_splice_site in intron
boundary loop to match behaviour in intron loop

In cases where there are multiple differing regions it was possible that $intron_effects->{splice_region}
could be set and then unset. Changed it so that once it's set it doesn't get unset.

modules/Bio/EnsEMBL/Variation/Utils/VariationEffect.pm
==
Fix to start_lost
Recent change
     return 0 unless $alt_pep && $alt_pep ne 'X';
Should be defined($alt_pep) rather than $alt_pep because deletion can mean $alt_pep is the empty string
Example: 438410 from clinvar

_overlaps_start_codon
Added some commented out code which would fix cases where _overlaps_start_codon misses overlaps.
This is commented out currently because other predicates rely on the existing incorrect behaviour.

Fix to _snp_start_altered
Fix to return values for checks for:
  isa('Bio::EnsEMBL::Variation::TranscriptStructuralVariationAllele')
  seq_is_unambiguous_dna();
  $cdna_start && $cdna_end
to return 1 rather than 0 so it's return value matches the default cache value that
would be returned in further calls to _snp_start_altered.

Returning 1 should be the default I think, only in cases where it is certain that the
start codon is "ATG" after the modification should 0 be returned.

Fix to _ins_del_start_altered
Fix to recent change to ins_del_start_altered
With a big deletion its possible that the length of utr_and_translateable is less than the length of the utr
leading to an error message from substr

Fix to stop_lost
Added a check for cds_end_NF. Can't be stop_lost if it isn't the stop codon, and if cds_end_NF
then the CDS is incomplete at the 3' end.

Fix to stop_retained
Replaced call to very buggy ref_eq_alt_sequence method with new stop_retained_checks which is
more restrictive about when stop_retained consequence is predicted. I'm still not happy with that
code, but as I don't know what the specific cases ref_eq_alt_sequence was trying to address I
have left the code as close to the original as I can.

modules/Bio/EnsEMBL/Variation/VariationFeatureOverlapAllele.pm
==
Fix to _get_differing_regions
Changed condition for when to split into multiple differing regions
 - Now only do on multibase substitutions where lengths are the same

This was to fix indel cases where some of the ranges generated falsely
indicated that splice regions near the end of introns were being affected by
the change (eg. an insert of 14 bases 20 bases from the end of a forward strand
intron does not extend to within 6 bases of the end of the intron, it pushes
the end of the intron 14 bases further away).

modules/t/test-genome-DBs/homo_sapiens/core/transcript_attrib.txt
==
Removed cds/rna_start/end_NF attribs where the attribute value was 0. The value
of 0 was indicating that the attribute was not set, ie it wasn't
a cds/rna_start/end_NF. It's not normal for the attrib to be
present when it has a value of 0, so removing these makes the test data
more representative of real data.

The VariationEffect.pm code assumes existence of one of these attributes indicated it was set,
which is a correct assumption for real EnsEMBL databases, but not for the test data.

variation_effect.t
==
test_count -> test_num and added various increments to actually count
all the tests done

Modified results of some tests affected by the change to the
test data (transcript_attrib table fixes).

Added a set of tests for a transcript with cds_start_NF and cds_end_NF
attributes added.

Fixed output so doesn't mess up TAP when there are failed tests
(was getting warnings that test numbers didn't match)

Removed die after failed pep_alleles test, replaced with a diag message

variationFeatureAdaptor.t
==
Removed 3 tests which were testing for the name of the first returned record
from fetch methods which did not use an ORDER BY in the sql. The tests
were expecting the data to be output in the same order it was inserted into
the table. With mysql 8 this was not true.

Added a test to check the correct number of vfs are returned by fetch_Iterator